feat(llm): T246.10 A6.1+A6.2 — batched prefill via linear-chain tree forward
Implement Qwen35ModelCudaQ4K::prefill_tokens(token_ids, start_pos) for all
four variants (Qwen2/Qwen3 PureTransformer, Dense hybrid, MoE hybrid). Uses
the existing tree-attention infrastructure (T246.7 P1.3 + TrackC.3) with a
degenerate linear-chain topology — parents = [-1, 0, 1, ..., N-2],
depths = [0, 1, ..., N-1] — and forces the post-forward acceptance walk to
accept all N BFS-ordered nodes (`force_accept_all = true`). The KV
compaction is a no-op for linear chains (src == i for every accepted slot)
and the counter advance writes all N positions in one go.

Changes
-------
- MAX_TREE_SIZE bumped 32 → 512 to accommodate pp512 prefill batches.
  Memory cost on Qwen3.6-35B-A3B MoE : ~6.6 GB extra (mostly SSM tree-state
  buffers for 48 SSM layers, decision 25db464b). Acceptable on GB10 256 GB.
- depths widened u8 → u16 (a chain of 512 has depth 511 which overflows
  u8). Rust + CUDA kernel signature both updated.
  - gqa_decode_tree_partial_bf16 : `const unsigned short* depths` (was u8)
  - DraftTree::depths : Vec<u16> (was Vec<u8>)
  - DraftTree::max_depth() : returns u16 (was u8)
  - delta_net_step_tree_bf16 unaffected (uses wave_indices, not depths)
- decode_step_tree_pure_transformer + decode_step_tree_hybrid renamed
  to *_inner with new `force_accept_all: bool` flag. Old public methods
  are now 1-line wrappers that pass false (verify semantics unchanged).
- decode_step_tree_hybrid_inner gains MoE FFN per-row dispatch
  (moe_ffn_forward_step_{mega,async,sync} called once per BFS row using
  the per-token MoE scratch buffers). This unlocks Qwen3.6-35B-A3B
  prefill ; the same path stays gated for MoE Lookahead via the
  decode_step_tree dispatcher (note d26271c4 — separate concern).
- Linear-chain memory footprint stays bounded : per-row attention/FFN
  scratch is reused via offsets ; SSM tree-state buffer is allocated
  once at model ctor (eager, per decision d148a083).

Acceptance path now :
  prefill_tokens(token_ids[..N], start_pos=self.position)
    → (validate, dispatch_by_variant)
    → decode_step_tree_*_inner(drafts=token_ids,
                               parents=chain,
                               depths=chain,
                               force_accept_all=true)
    → returns argmax[N-1] = the N+1-th token prediction

N=1 fast-path : delegates to decode_step (preserves captured graph win).

Quality gates (local, non-CUDA)
- cargo check / build / clippy / fmt clean for rustorch-llm.
- cargo test -p rustorch-llm : 36/36 + 1/1 lookahead tests pass.
- depth widening : existing parity test (cuda_tree_parity, [0u8] → [0u16])
  + smoke tests (qwen36_cuda_decode_step_tree_smoke, qwen36_hybrid_tree_smoke)
  updated mechanically.

CUDA build / DGX bench / parity test ship in follow-up commit.

+++ crates/rustorch-llm/src/lookahead/jacobi.rs
@@ -40,8 +40,10 @@ pub struct DraftTree {
     /// `parents[i]` is the index of node `i`'s parent, or `-1` for
     /// the root. Always `< i` for non-root nodes (BFS invariant).
     pub parents: Vec<i32>,
-    /// Depth of each node from the root (root depth = 0).
-    pub depths: Vec<u8>,
+    /// Depth of each node from the root (root depth = 0). T246.10 A6
+    /// widened to u16 so prefill linear-chain trees of length > 256
+    /// remain representable.
+    pub depths: Vec<u16>,
 }
 
 impl DraftTree {
@@ -57,7 +59,7 @@ impl DraftTree {
     }
 
     /// Maximum depth observed (root depth = 0).
-    pub fn max_depth(&self) -> u8 {
+    pub fn max_depth(&self) -> u16 {
         self.depths.iter().copied().max().unwrap_or(0)
     }
 
@@ -300,7 +302,7 @@ impl JacobiWindow {
                     let new_idx = tree.tokens.len();
                     tree.tokens.push(t);
                     tree.parents.push(parent_idx as i32);
-                    tree.depths.push(d as u8);
+                    tree.depths.push(d as u16);
                     cur_level_indices.push(new_idx);
                 }
             }
